Extract RequestContextDecorator from the PublicResource
