+++ src/main/java/edu/illinois/library/cantaloupe/resource/RequestContextDecorator.java
@@ -1,0 +1,60 @@
+package edu.illinois.library.cantaloupe.resource;
+
+import java.util.Enumeration;
+
+import edu.illinois.library.cantaloupe.delegate.DelegateProxyService;
+import edu.illinois.library.cantaloupe.http.Cookies;
+import edu.illinois.library.cantaloupe.image.MetaIdentifier;
+import edu.illinois.library.cantaloupe.image.ScaleConstraint;
+import edu.illinois.library.cantaloupe.http.Reference;
+
+public class RequestContextDecorator {
+    public static void decorateRequestContext(RequestContext context, MetaIdentifier metaID, Reference requestURI, Request request) {
+        if (!DelegateProxyService.isDelegateAvailable()) {
+            return;
+        }
+        context.setLocalURI(request.getReference());
+        context.setRequestURI(requestURI);
+        context.setRequestHeaders(request.getHeaders().toMap());
+        context.setClientIP(getCanonicalClientIPAddress(request));
+        context.setCookies(getCookies(request).toMap());
+        if (metaID != null) {
+            context.setIdentifier(metaID.getIdentifier());
+            context.setPageNumber(metaID.getPageNumber());
+            ScaleConstraint scaleConstraint = metaID.getScaleConstraint();
+            if (scaleConstraint == null) {
+                // Delegate users will appreciate not having to check for
+                // null.
+                scaleConstraint = new ScaleConstraint(1, 1);
+            }
+            context.setScaleConstraint(scaleConstraint);
+        }
+    }
+
+        /**
+     * @return User agent's IP address, respecting the {@code X-Forwarded-For}
+     *         request header, if present.
+     */
+    private static String getCanonicalClientIPAddress(Request request) {
+        // The value is expected to be in the format: "client, proxy1, proxy2"
+        final String forwardedFor =
+                request.getHeaders().getFirstValue("X-Forwarded-For", "");
+        if (!forwardedFor.isEmpty()) {
+            return forwardedFor.split(",")[0].trim();
+        } else {
+            // Fall back to the client IP address.
+            return request.getRemoteAddr();
+        }
+    }
+
+    public static Cookies getCookies(Request request) {
+        Cookies cookies = new Cookies();
+            final Enumeration<String> headers = request.getServletRequest().getHeaders("Cookie");
+            while (headers.hasMoreElements()) {
+                String value = headers.nextElement();
+                Cookies batch = Cookies.fromHeaderValue(value);
+                cookies.addAll(batch);
+            }
+        return cookies;
+    }
+}
