+++ src/main/java/edu/illinois/library/cantaloupe/resource/iiif/IIIFResource.java
@@ -2,7 +2,6 @@ package edu.illinois.library.cantaloupe.resource.iiif;
 
 import edu.illinois.library.cantaloupe.config.Configuration;
 import edu.illinois.library.cantaloupe.config.Key;
-import edu.illinois.library.cantaloupe.delegate.DelegateProxyService;
 import edu.illinois.library.cantaloupe.http.Reference;
 import edu.illinois.library.cantaloupe.http.Status;
 import edu.illinois.library.cantaloupe.image.Dimension;
@@ -15,7 +14,7 @@ import edu.illinois.library.cantaloupe.operation.OperationList;
 import edu.illinois.library.cantaloupe.operation.Scale;
 import edu.illinois.library.cantaloupe.operation.ScaleByPixels;
 import edu.illinois.library.cantaloupe.resource.AbstractResource;
-import edu.illinois.library.cantaloupe.resource.RequestContext;
+import edu.illinois.library.cantaloupe.resource.RequestContextDecorator;
 import edu.illinois.library.cantaloupe.resource.ScaleRestrictedException;
 import edu.illinois.library.cantaloupe.resource.StringRepresentation;
 import edu.illinois.library.cantaloupe.util.TimeUtils;
@@ -44,26 +43,11 @@ public abstract class IIIFResource extends AbstractResource {
     @Override
     public void doInit() throws Exception {
         super.doInit();
-        if (DelegateProxyService.isDelegateAvailable()) {
-            RequestContext context = getRequestContext();
-            context.setLocalURI(getRequest().getReference());
-            context.setRequestURI(getPublicReference());
-            context.setRequestHeaders(getRequest().getHeaders().toMap());
-            context.setClientIP(getCanonicalClientIPAddress());
-            context.setCookies(getRequest().getCookies().toMap());
-            MetaIdentifier metaID = getMetaIdentifier();
-            if (metaID != null) {
-                context.setIdentifier(metaID.getIdentifier());
-                context.setPageNumber(metaID.getPageNumber());
-                ScaleConstraint scaleConstraint = metaID.getScaleConstraint();
-                if (scaleConstraint == null) {
-                    // Delegate users will appreciate not having to check for
-                    // null.
-                    scaleConstraint = new ScaleConstraint(1, 1);
-                }
-                context.setScaleConstraint(scaleConstraint);
-            }
-        }
+        RequestContextDecorator.decorateRequestContext(
+                            getRequestContext(),
+                            getMetaIdentifier(),
+                            getPublicReference(),
+                            getRequest());
         addHeaders();
     }
 
@@ -106,22 +90,6 @@ public abstract class IIIFResource extends AbstractResource {
                 getResponse().setHeader("Cache-Control",
                         String.join(", ", directives));
             }
-        }
-    }
-
-    /**
-     * @return User agent's IP address, respecting the {@code X-Forwarded-For}
-     *         request header, if present.
-     */
-    private String getCanonicalClientIPAddress() {
-        // The value is expected to be in the format: "client, proxy1, proxy2"
-        final String forwardedFor =
-                getRequest().getHeaders().getFirstValue("X-Forwarded-For", "");
-        if (!forwardedFor.isEmpty()) {
-            return forwardedFor.split(",")[0].trim();
-        } else {
-            // Fall back to the client IP address.
-            return getRequest().getRemoteAddr();
         }
     }
